Move generate_column_map into the __init__.py module. Get data from the constant instead of the DB Accessor. Move the test_generate_column_map out of the test db accessor file and into the reporting_common directory.
