+++ koku/masu/processor/aws/aws_report_summary_updater.py
@@ -22,10 +22,10 @@ import logging
 from tenant_schemas.utils import schema_context
 
 from masu.database.aws_report_db_accessor import AWSReportDBAccessor
-from masu.database.reporting_common_db_accessor import ReportingCommonDBAccessor
 from masu.external.date_accessor import DateAccessor
 from masu.util.aws.common import get_bills_from_provider
 from masu.util.common import date_range_pair
+from reporting_common import REPORT_COLUMN_MAP
 
 LOG = logging.getLogger(__name__)
 
@@ -44,8 +44,7 @@ class AWSReportSummaryUpdater:
         self._schema = schema
         self._provider = provider
         self._manifest = manifest
-        with ReportingCommonDBAccessor() as reporting_common:
-            self._column_map = reporting_common.column_map
+        self._column_map = REPORT_COLUMN_MAP
         self._date_accessor = DateAccessor()
 
     def update_daily_tables(self, start_date, end_date):
