+++ koku/masu/processor/azure/azure_report_summary_updater.py
@@ -22,10 +22,10 @@ import logging
 from tenant_schemas.utils import schema_context
 
 from masu.database.azure_report_db_accessor import AzureReportDBAccessor
-from masu.database.reporting_common_db_accessor import ReportingCommonDBAccessor
 from masu.external.date_accessor import DateAccessor
 from masu.util.azure.common import get_bills_from_provider
 from masu.util.common import date_range_pair
+from reporting_common import REPORT_COLUMN_MAP
 
 LOG = logging.getLogger(__name__)
 
@@ -44,8 +44,7 @@ class AzureReportSummaryUpdater:
         self._schema = schema
         self._provider = provider
         self._manifest = manifest
-        with ReportingCommonDBAccessor() as reporting_common:
-            self._column_map = reporting_common.column_map
+        self._column_map = REPORT_COLUMN_MAP
         self._date_accessor = DateAccessor()
 
     def _get_sql_inputs(self, start_date, end_date):
