+++ koku/masu/processor/ocp/ocp_cloud_updater_base.py
@@ -21,8 +21,8 @@ import logging
 from api.provider.models import Provider
 from masu.database.ocp_report_db_accessor import OCPReportDBAccessor
 from masu.database.provider_db_accessor import ProviderDBAccessor
-from masu.database.reporting_common_db_accessor import ReportingCommonDBAccessor
 from masu.external.date_accessor import DateAccessor
+from reporting_common import REPORT_COLUMN_MAP
 
 LOG = logging.getLogger(__name__)
 
@@ -46,8 +46,7 @@ class OCPCloudUpdaterBase:
         self._provider = provider
         self._provider_uuid = str(self._provider.uuid)
         self._manifest = manifest
-        with ReportingCommonDBAccessor() as reporting_common:
-            self._column_map = reporting_common.column_map
+        self._column_map = REPORT_COLUMN_MAP
         self._date_accessor = DateAccessor()
 
     def get_infra_map(self):
