+++ koku/masu/test/celery/test_tasks.py
@@ -27,10 +27,10 @@ from api.utils import DateHelper
 from masu.celery import tasks
 from masu.celery.export import TableExportSetting
 from masu.database.report_manifest_db_accessor import ReportManifestDBAccessor
-from masu.database.reporting_common_db_accessor import ReportingCommonDBAccessor
 from masu.test import MasuTestCase
 from masu.test.database.helpers import ReportObjectCreator
 from masu.util.common import dictify_table_export_settings
+from reporting_common import REPORT_COLUMN_MAP
 
 fake = faker.Faker()
 DummyS3Object = namedtuple("DummyS3Object", "key")
@@ -345,8 +345,7 @@ class TestUploadTaskWithData(MasuTestCase):
         """Set up initial data for tests."""
         super().setUp()
 
-        with ReportingCommonDBAccessor(self.schema) as common_accessor:
-            self.column_map = common_accessor.column_map
+        self.column_map = REPORT_COLUMN_MAP
         self.creator = ReportObjectCreator(self.schema, self.column_map)
 
         timezone = pytz.timezone("UTC")
