+++ koku/masu/test/database/test_ocp_report_db_accessor.py
@@ -30,7 +30,6 @@ from api.utils import DateHelper
 from masu.database import OCP_REPORT_TABLE_MAP
 from masu.database.ocp_report_db_accessor import OCPReportDBAccessor
 from masu.database.provider_db_accessor import ProviderDBAccessor
-from masu.database.reporting_common_db_accessor import ReportingCommonDBAccessor
 from masu.external.date_accessor import DateAccessor
 from masu.test import MasuTestCase
 from masu.test.database.helpers import ReportObjectCreator
@@ -39,6 +38,7 @@ from reporting.models import OCPUsageLineItem
 from reporting.models import OCPUsageLineItemDailySummary
 from reporting.models import OCPUsageReport
 from reporting.models import OCPUsageReportPeriod
+from reporting_common import REPORT_COLUMN_MAP
 
 
 class OCPReportDBAccessorTest(MasuTestCase):
@@ -49,8 +49,7 @@ class OCPReportDBAccessorTest(MasuTestCase):
         """Set up the test class with required objects."""
         super().setUpClass()
 
-        cls.common_accessor = ReportingCommonDBAccessor()
-        cls.column_map = cls.common_accessor.column_map
+        cls.column_map = REPORT_COLUMN_MAP
         cls.accessor = OCPReportDBAccessor(schema=cls.schema, column_map=cls.column_map)
         cls.report_schema = cls.accessor.report_schema
         cls.creator = ReportObjectCreator(cls.schema, cls.column_map)
