+++ koku/masu/test/database/test_reporting_common_db_accessor.py
@@ -15,7 +15,6 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 """Test the ReportingCommonDBAccessor utility object."""
-import copy
 from unittest.mock import Mock
 
 from masu.database import AWS_CUR_TABLE_MAP
@@ -35,22 +34,6 @@ class ReportingCommonDBAccessorTest(MasuTestCase):
     def test_initializer(self):
         """Test initializer."""
         self.assertIsInstance(self.accessor.column_map, dict)
-
-    def test_generate_column_map(self):
-        """Assert all tables are in the column map."""
-        column_map = self.accessor.generate_column_map()
-        keys = column_map.keys()
-
-        tables = copy.deepcopy(self.report_tables)
-        tables.remove(AWS_CUR_TABLE_MAP["cost_entry"])
-        tables.remove(AWS_CUR_TABLE_MAP["line_item_daily"])
-        tables.remove(AWS_CUR_TABLE_MAP["line_item_daily_summary"])
-        tables.remove(AWS_CUR_TABLE_MAP["tags_summary"])
-        tables.remove(AWS_CUR_TABLE_MAP["ocp_on_aws_daily_summary"])
-        tables.remove(AWS_CUR_TABLE_MAP["ocp_on_aws_project_daily_summary"])
-        tables.remove(AWS_CUR_TABLE_MAP["ocp_on_aws_tags_summary"])
-        for table in tables:
-            self.assertIn(table, keys)
 
     def test_add(self):
         """Test the add() function."""
