+++ koku/masu/test/processor/aws/test_aws_report_db_cleaner.py
@@ -22,11 +22,11 @@ from tenant_schemas.utils import schema_context
 
 from masu.database import AWS_CUR_TABLE_MAP
 from masu.database.aws_report_db_accessor import AWSReportDBAccessor
-from masu.database.reporting_common_db_accessor import ReportingCommonDBAccessor
 from masu.processor.aws.aws_report_db_cleaner import AWSReportDBCleaner
 from masu.processor.aws.aws_report_db_cleaner import AWSReportDBCleanerError
 from masu.test import MasuTestCase
 from masu.test.database.helpers import ReportObjectCreator
+from reporting_common import REPORT_COLUMN_MAP
 
 
 class AWSReportDBCleanerTest(MasuTestCase):
@@ -36,8 +36,7 @@ class AWSReportDBCleanerTest(MasuTestCase):
     def setUpClass(cls):
         """Set up the test class with required objects."""
         super().setUpClass()
-        cls.common_accessor = ReportingCommonDBAccessor()
-        cls.column_map = cls.common_accessor.column_map
+        cls.column_map = REPORT_COLUMN_MAP
         cls.accessor = AWSReportDBAccessor(schema=cls.schema, column_map=cls.column_map)
         cls.report_schema = cls.accessor.report_schema
         cls.creator = ReportObjectCreator(cls.schema, cls.column_map)
