+++ koku/masu/test/processor/azure/test_azure_report_charge_updater.py
@@ -16,10 +16,10 @@
 #
 """Test the AzureCostModelCostUpdater object."""
 from masu.database.azure_report_db_accessor import AzureReportDBAccessor
-from masu.database.reporting_common_db_accessor import ReportingCommonDBAccessor
 from masu.external.date_accessor import DateAccessor
 from masu.processor.azure.azure_cost_model_cost_updater import AzureCostModelCostUpdater
 from masu.test import MasuTestCase
+from reporting_common import REPORT_COLUMN_MAP
 
 
 class AzureCostModelCostUpdaterTest(MasuTestCase):
@@ -33,8 +33,7 @@ class AzureCostModelCostUpdaterTest(MasuTestCase):
 
         updater.update_summary_cost_model_costs()
 
-        with ReportingCommonDBAccessor() as report_common_db:
-            column_map = report_common_db.column_map
+        column_map = REPORT_COLUMN_MAP
 
         with AzureReportDBAccessor(self.schema, column_map) as accessor:
             bill = accessor.get_cost_entry_bills_by_date(bill_date)[0]
