+++ koku/masu/test/processor/gcp/test_gcp_report_processor.py
@@ -17,7 +17,6 @@ from api.provider.models import ProviderAuthentication
 from api.provider.models import ProviderBillingSource
 from masu.database.gcp_report_db_accessor import GCPReportDBAccessor
 from masu.database.report_manifest_db_accessor import ReportManifestDBAccessor
-from masu.database.reporting_common_db_accessor import ReportingCommonDBAccessor
 from masu.external import UNCOMPRESSED
 from masu.external.date_accessor import DateAccessor
 from masu.processor.gcp.gcp_report_processor import GCPReportProcessor
@@ -26,6 +25,7 @@ from masu.util import common as utils
 from reporting.provider.gcp.models import GCPCostEntryBill
 from reporting.provider.gcp.models import GCPCostEntryLineItemDaily
 from reporting.provider.gcp.models import GCPProject
+from reporting_common import REPORT_COLUMN_MAP
 
 fake = Faker()
 
@@ -42,8 +42,7 @@ class GCPReportProcessorTest(MasuTestCase):
         cls.date_accessor = DateAccessor()
         cls.manifest_accessor = ReportManifestDBAccessor()
 
-        with ReportingCommonDBAccessor() as report_common_db:
-            cls.column_map = report_common_db.column_map
+        cls.column_map = REPORT_COLUMN_MAP
 
     def setUp(self):
         """Set up GCP tests."""
